corrections to and add google analytics to main layout
