+++ src/app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from "next";
 import { Playfair_Display, Open_Sans } from "next/font/google";
 import "../styles/styles.scss";
-
 import { Header, Footer } from "@/components/navigation";
+import { GoogleAnalytics } from "@/components/analytics/GoogleAnalytics";
+import Script from "next/script";
 
 const playfairDisplay = Playfair_Display({
   variable: "--font-playfair-display",
@@ -37,9 +38,21 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
+      <head>
+        <Script src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_ID}`} strategy="afterInteractive" />
+        <Script id="gtag-init" strategy="afterInteractive">
+          {`
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', '${process.env.NEXT_PUBLIC_GA_ID}', {
+              page_path: window.location.pathname,
+            });
+          `}
+        </Script>
+      </head>
       <body className={`${playfairDisplay.variable} ${openSans.variable}`}>
-        {" "}
-        <Header />
+        <GoogleAnalytics /> <Header />
         {children}
         <Footer />
       </body>
